requirements.txt
pandas
requests
user-agents
maxminddb-geolite2
pandarallel
click
flask
pytest
pytest-cov
